update tests1 for new bounds
