--- test/scenario_2/stations1.csv
@@ -1,2 +1,0 @@
-ID,Lat,Lon,Hgt_m
-MHMS,33.939,-118.244,0.1
