+++ test/scenario_2/stations2.csv
@@ -1,0 +1,4 @@
+ID,Lat,Lon,Hgt_m
+PIN1,33.61216,-116.45816,1256.16944
+TORP,33.79780,-118.33060,-5.21643
+
