--- test/scenario_3/S1_sv_file.txt
@@ -1,8 +1,0 @@
-2018-11-12T23:00:02.000000 -2064965.285362 6434865.494987 2090670.967443 860.239634 2590.964968 -7090.378144
-2018-11-12T23:00:12.000000 -2056228.553736 6460407.492520 2019650.417312 887.072466 2517.380329 -7113.598127
-2018-11-12T23:00:22.000000 -2047224.526705 6485212.031660 1948401.684024 913.698134 2443.474728 -7136.014344
-2018-11-12T23:00:32.000000 -2037955.293282 6509275.946120 1876932.818066 940.113169 2369.256838 -7157.624244
-2018-11-12T23:00:42.000000 -2028422.977002 6532596.156540 1805251.894958 966.314136 2294.735374 -7178.425371
-2018-11-12T23:00:52.000000 -2018629.735564 6555169.670917 1733367.014327 992.297636 2219.919093 -7198.415359
-2018-11-12T23:01:02.000000 -2008577.760461 6576993.585012 1661286.298987 1018.060311 2144.816789 -7217.591940
-2018-11-12T23:01:12.000000 -1998269.276601 6598065.082739 1589017.893976 1043.598837 2069.437298 -7235.952940
