--- test/scenario_3/incorrect_file.txt
@@ -1,1 +1,0 @@
-this is a dummy file
